\name{obesity}
\alias{obesity}
\title{
Team Package
}
\description{
This is a package for team
}
\usage{
obesity(X, Y)
}

\arguments{
  \item{X}{
ddd
}
  \item{Y}{
ddd
}
}
\details{
ddd
}
\value{
dddd
}
\references{
ddd
}
\author{
Zhendan Zhu
}
\note{
ddd
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (X, Y) 
{
    Z = X^2 + Y^2
  }
}
