\name{teamrpack-package}
\alias{teamrpack-package}
\docType{package}
\title{teamrpack: Team R Package}
\description{teamrpack: Team R Package}
\details{This package contains crowdsim.R and randomnums.RData. crowdsim.R is a crowd simulator and randomnums.RData is a set of 100 random numbers drawn from a normal distribution. See included individual .Rd files for more information.}
\author{Steven Su <ssu7@jhu.edu>}
\keyword{package}
\examples{
crowdsim(2,100, 10, 10, 10)
crowdsim(3,10, 10, 11, 1)
data(randomnums)
}
